plugged in fontawesome icons renamed action files
