+++ src/components/Header.js
@@ -21,7 +21,7 @@ class Header extends Component {
         <section className="header-container">
           <div className="logo">
             <img src={logo}/>
-            <h1>Categories</h1>
+            <h1>Expenses by Budget Category</h1>
           </div>
           <nav>
             <ul>
