+++ src/components/category-actions.js
@@ -4,7 +4,7 @@ import {
   CATEGORY_UPDATE,
   CATEGORY_REMOVE 
 } from './reducers.js';
-import data from './categories/categories-data';
+import data from './categories/categories-data.js';
 import shortid from 'shortid';
 
 export const load = () => ({
